GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch (refs/gitbutler/Virtual-branch)
 - Update python dependencies (refs/gitbutler/Update-python-dependencies)
   - package/bin/util.py
   - tests/test_util.py
   - requirements_dev.txt
   - package/lib/requirements.txt
 - add-proxy-support-to-apiv2 (refs/gitbutler/add-proxy-support-to-apiv2)
   branch head: 42a57fe5ebf3acb442239aef70d616099f0dc1b3
   - tests/test_util.py
   - package/bin/util.py
   - package/bin/dynatrace_api_v2.py
   - package/bin/dynatrace_problem.py
   - package/bin/dynatrace_entity.py
   - package/bin/dynatrace_timeseries_metrics.py
   - package/bin/dynatrace_timeseries_single_metric.py
   - package/bin/dynatrace_timeseries_metrics_v2.py
   - .idea/misc.xml
 - Update version and add ucc version (refs/gitbutler/Update-version-and-add-ucc-version)
   - globalConfig.json

Your previous branch was: refs/heads/add-proxy-support-to-apiv2

The sha for that commit was: 42a57fe5ebf3acb442239aef70d616099f0dc1b3

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch